+++ crates/miden-protocol/src/testing/note.rs
@@ -1,11 +1,11 @@
 use alloc::vec::Vec;
 
+use crate::Word;
 use crate::assembly::Assembler;
 use crate::asset::FungibleAsset;
 use crate::note::{
     Note,
     NoteAssets,
-    NoteExecutionHint,
     NoteInputs,
     NoteMetadata,
     NoteRecipient,
@@ -14,7 +14,6 @@ use crate::note::{
     NoteType,
 };
 use crate::testing::account_id::ACCOUNT_ID_SENDER;
-use crate::{Word, ZERO};
 
 pub const DEFAULT_NOTE_CODE: &str = "begin nop end";
 
@@ -29,8 +28,6 @@ impl Note {
             sender_id,
             NoteType::Private,
             NoteTag::with_account_target(sender_id),
-            NoteExecutionHint::Always,
-            ZERO,
         );
         let inputs = NoteInputs::new(Vec::new()).unwrap();
         let recipient = NoteRecipient::new(serial_num, note_script, inputs);